Cleaned up stats
I did this because the description for the RC24-supported channels were old and that's when Spotlight and the others were considering writing them in Go. Not much happened from that. I left them as considering in case we do a rewrite of the code in another language, something I'm not willing to do since I like it in Python. Talk to me on Discord if you have issues with what I did.